requirements.txt
elevenlabs~=0.2.24
flake8
openai~=0.27.8
openai-whisper
pyaudio~=0.2.13
pynput~=1.7.6
python-decouple
pyyaml~=6.0.1
flask~=3.0.0

pygame~=2.5.2
requests~=2.31.0
twitchAPI~=4.1.0
numpy~=1.24.4
audioread~=3.0.0